+++ src/ui/simulator_backend.h
@@ -27,17 +27,17 @@ public:
     void createSprite(int16_t w, int16_t h) override;
     int16_t width() override;
     int16_t height() override;
-    void fillSprite(uint16_t color) override;
-    void drawLine(int16_t x1, int16_t y1, int16_t x2, int16_t y2, uint16_t color) override;
-    void drawRect(int16_t x, int16_t y, int16_t w, int16_t h, uint16_t color) override;
-    void fillRect(int16_t x, int16_t y, int16_t w, int16_t h, uint16_t color) override;
-    void drawFastHLine(int16_t x, int16_t y, int16_t w, uint16_t color) override;
+    void fillSprite(uint32_t color) override;
+    void drawLine(int16_t x1, int16_t y1, int16_t x2, int16_t y2, uint32_t color) override;
+    void drawRect(int16_t x, int16_t y, int16_t w, int16_t h, uint32_t color) override;
+    void fillRect(int16_t x, int16_t y, int16_t w, int16_t h, uint32_t color) override;
+    void drawFastHLine(int16_t x, int16_t y, int16_t w, uint32_t color) override;
     void drawString(const char* str, int16_t x, int16_t y) override;
     void drawRightString(const char* str, int16_t x, int16_t y) override;
     void drawCenterString(const char* str, int16_t x, int16_t y) override;
     void setFont(const void* font) override;
-    void setTextColor(uint16_t fg, uint16_t bg) override;
-    uint16_t color565(uint8_t r, uint8_t g, uint8_t b) override;
+    void setTextColor(uint32_t fg, uint32_t bg) override;
+    uint32_t color565(uint8_t r, uint8_t g, uint8_t b) override;
     void pushSprite(int16_t x, int16_t y) override;
     const char* getCurrentFontName() override; // Ensure this matches CanvasWrapper
     void* getCanvas() override { return this; }
@@ -47,7 +47,7 @@ public:
     TTF_Font* getTTFFont() { return ttfFont; }
 
 private:
-    void setPixel(int32_t x, int32_t y, uint16_t color);  // Helper method
+    void setPixel(int32_t x, int32_t y, uint32_t color);  // Helper method
 };
 
 #endif  // SIMULATOR
